Completed Module 4

+++ Camedia/Models/Genre.cs
@@ -7,7 +7,7 @@ namespace Camedia.Models
 {
 	public class Genre
 	{
-		public int Id { get; set; }
+		public Byte Id { get; set; }
 
 		public string Name { get; set; }
 	}
